🔥 refactor(home_history_and_submission.dart): remove unused imports and code This commit removes unused imports and code from the home_history_and_submission.dart file. This improves the code's readability and reduces the file size.
🔨 refactor(home_history_and_submission.dart): extract ListView to a separate widget
The ListView widget is extracted to a separate widget named ListRiwayatPengajuan. This improves the readability and maintainability of the code.

🚀 feat(home_history_and_submission.dart): add ListRiwayatPengajuan widget
This commit adds a new widget called ListRiwayatPengajuan to the HomeHistoryAndSubmission class. The widget displays a list of submissions made by the user and includes a refresh indicator for refreshing the list. The widget also includes a ListTile for each submission that displays the name of the borrower, the submission ID, and a random avatar.