🎨 style(home_history_and_submission.dart): refactor TitlePengajuan into a separate widget This commit refactors the TitlePengajuan widget into a separate widget to improve code readability and maintainability. The new widget takes in the necessary parameters and returns the same UI as before. No functionality has been changed.
🎨 style(home_history_and_submission.dart): refactor ListView to include Scrollbar
This commit refactors the ListView to include a Scrollbar widget. The Scrollbar widget provides a visual indication of the current position within the list and makes it easier for users to navigate through the list.

📦 chore(pubspec.yaml): add bs_flutter_alert package
The bs_flutter_alert package has been added to the dependencies list in the pubspec.yaml file. This package will be used to display alerts in the application.