🎨 style(search.png): add search icon to main page The search icon has been added to the main page to improve the user experience by making it easier to find and use the search functionality.
🔥 refactor(home_search_nik.dart): remove unused import statement
The import statement for the Lottie package is no longer used in the file and has been removed to improve code readability and maintainability.

🎨 style(home_search_nik.dart): refactor HomeSearchNik widget
This commit refactors the HomeSearchNik widget to improve its readability and maintainability. The changes include replacing the ClipRRect widget with a Card widget with a rounded border, adding a background image to the Card widget, and separating the NikSearchForm and NikSearchResult widgets into their own classes. The code has also been formatted to conform to the Dart style guide.

🚀 chore(pubspec.yaml): add new image directory to assets
Added a new directory to the assets section of the pubspec.yaml file. The new directory contains images that will be used in the main page of the home section of the application.