🎨 style(login_page_view.dart): refactor LoginPageView widget The LoginPageView widget has been refactored to improve readability and maintainability. The WavesLogin widget has been extracted into its own file for better separation of concerns. The LoginForm widget has also been extracted into its own file to improve code organization.
✨ feat(login-page): add login form and waves login widget
The login form widget contains a form with email and password fields, with validation and custom styling. The waves login widget is a custom widget that displays a wave animation in the background of the login page. It takes in a list of colors, durations, and height percentages to customize the wave animation.

🔨 chore(pubspec.yaml): update dependencies
This commit updates the dependencies in the pubspec.yaml file. The following dependencies were updated:
- data_table_2: ^2.4.2
- device_info_plus: ^8.2.0
- dio: ^5.1.1
- firebase_analytics: ^10.2.0
- firebase_auth: ^4.4.0
- firebase_core: ^2.9.0
- firebase_crashlytics: ^3.1.0
- firebase_messaging: ^14.4.0
- flutter_map_marker_popup: ^4.0.2
- flutter_svg: ^2.0.5
- google_sign_in: ^6.1.0
- image_picker: ^0.8.7+2
- lottie: ^2.3.2
- share_plus: ^6.3.2
-