✨ feat(login-page): add login page components This commit adds four new files to the login-page module. The Attribution component displays the logos of two regulatory bodies and a call center. The NotPlannedGoogleSignIn component is a Google sign-in button that is currently not planned for use. The LoginButton component is a button that triggers the login process. The RegisterButton component is a button that displays a dialog box with a message to contact the admin to register an account.
✨ feat(login_form.dart): add FleetimeLoginTextField widget
This commit adds a new widget called FleetimeLoginTextField to the login form. The widget is a reusable text field that can be used to collect user input. It has several properties such as name, controller, icon, hintText, validator, obscureText, and suffixIcon. These properties allow for customization of the text field to suit different use cases. The widget is built using the FlutterFormBuilderTextField widget and has custom styling using the GoogleFonts package.

🎨 style(login_page_login_form.dart): reformat code and remove unused imports
This commit reformats the code to improve readability and removes unused imports to reduce clutter. No functionality has been changed.

🎨 style(login_page_login_form.dart): refactor login form with FleetimeLoginTextField
The login form has been refactored to use the custom FleetimeLoginTextField widget instead of the FormBuilderTextField widget. The FleetimeLoginTextField widget provides a more consistent and reusable design for the login form. The widget also includes support for icons, hint text, and password visibility.

🔥 refactor(login_page_login_form.dart): remove unused code
The code for the login button and the attribution widget have been moved to their own separate widgets, LoginButton and Attribution, respectively. The code for the Google sign-in button has been moved to a separate widget, NotPlannedGoogleSignIn. The code for the removed widgets has been deleted.