🔥 refactor(office_selection_view.dart): extract office selection form to a separate widget The office selection form was extracted to a separate widget called OfficeSelectionOfficeSearchableForm. This improves the readability and maintainability of the code. The extracted widget is now used in the OfficeSelectionView.
🎉 feat(searchabledropdown.dart): add FleetimeOfficeSelectionDropdown widget
🎉 feat(office_selection_office_searchable_form.dart): add OfficeSearchableBody widget
The FleetimeOfficeSelectionDropdown widget is added to provide a searchable dropdown for office selection. The OfficeSearchableBody widget is added to provide a column of FleetimeOfficeSelectionDropdown widgets for selecting the main and assistant office branches. This improves the user experience by providing a more intuitive and user-friendly way of selecting office branches.