🎨 style(home): add list_pending.png image asset The commit adds a new image asset to the home directory. The image is named list_pending.png and is used to display a list of pending items.
✨ feat(list_pending.dart): add new widget for displaying pending list
This commit adds a new widget for displaying a list of pending items for a reviewer. The widget includes a refresh indicator, a scrollbar, and a list view builder. Each item in the list includes a card with a leading avatar, a title, and a table of details. The title includes the name of the borrower, and the table includes the application number, submission date, analyst name, and status. The status is displayed as a chip with different colors based on the status value. The widget also includes a dialog box that appears when an item is clicked. The dialog box displays the borrower's name and provides two buttons for reviewing the application or viewing the progress.

🎨 style(reviewer_pending_list_available.dart): reformat code and remove unused imports
This commit reformats the code to improve readability and removes unused imports. No functionality has been changed.

🎨 style(reviewer_pending_list_available.dart): refactor ReviewerPendingList widget layout
The ReviewerPendingList widget layout has been refactored to improve its visual appearance. The RefreshIndicator widget has been removed and replaced with a Stack widget that contains an Image.asset widget and a Column widget. The Column widget contains a BootstrapAlert widget and a ListPendingPengajuanReviewer widget. The ListPendingPengajuanReviewer widget has been wrapped in an Expanded widget to ensure that it takes up all the remaining space in the Column widget. The BootstrapAlert widget has been given a margin of 10 and an onDismissed callback function.