🎨 style(home_controller.dart, pengutus_pending_list_view.dart): refactor dialog widgets The AwesomeDialog widgets in the home_controller.dart file have been refactored to SuccessDialog and ErrorDialog widgets for better readability and maintainability. In the pengutus_pending_list_view.dart file, the code has been refactored to use custom widgets PengutusEmptyPendingList and PengutusPendingList for better code organization and readability.
✨ feat(pengutus_pending_list): add new widget for available pengutus pending list
This commit adds a new widget for displaying the available pengutus pending list. The widget includes a refresh indicator, a scrollbar, and a list of pending pemutusan pengutus. The list displays the debtor's name, the submission date, the analyst's name, and the status of the submission. The widget also includes a BootstrapAlert that reminds the user to review the pending submissions.

🎨 style(pengutus_pending_list_empty.dart): fix super constructor call
🚀 feat(dialog_box.dart): add SuccessDialog class
🚀 feat(simple_alert.dart): add NativePromptAlertPink class
The super constructor call in PengutusEmptyPendingList was fixed to include the 'key' parameter. SuccessDialog class was added to dialog_box.dart to display a success dialog box. NativePromptAlertPink class was added to simple_alert.dart to display a pink alert dialog box with two buttons.