🐛 fix(server.ts): change port variable case from lowercase port to uppercase PORT ✨ feat(server.ts): add support for process.env.PORT environment variable The port variable is now named PORT, which improves consistency with the naming conventions as PORT is a constant. Support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.
🎨 style(home_controller.dart, useer_progression.dart, home_main_page.dart, home_profile.dart, rugi_laba_export.dart, edit_rugi_laba_view.dart): format code
This commit formats the code in the mentioned files to improve readability and maintainability.

✨ feat(home_controller.dart, useer_progression.dart): add targetPlafond and successPercentage variables
This commit adds targetPlafond and successPercentage variables to the HomeController and UserStatAndProgression classes, respectively. targetPlafond is a MoneyMask

🌐 chore(rugi_laba_view.dart): add locale to date format
The date format in the RugiLabaView and LihatRugiLabaView widgets now includes a locale parameter to format the date in Indonesian. This improves the user experience for Indonesian users who are more familiar with the Indonesian date format.